3D描画関数の整備。

+++ Update_Yoto/Update_Yoto/3DLibrary/3DLibrary/Scene/TitleScene.cpp
@@ -9,7 +9,7 @@ void TitleScene::Init()
 	m_State = SceneState::eMain;
 
 	cTexture::GetTextureInstance()->LoadTexture("3DLibrary/3DLibrary/Res/title.png",
-		(int)TextureCategory::TitleCategory,
+		TextureCategory::TitleCategory,
 		(int)TitleCategoryTextureList::Background);
 }
 
@@ -54,6 +54,6 @@ void TitleScene::Draw()
 		0.0f,
 		cTexture::GetTextureInstance()
 		->GetTexture(
-		(int)TextureCategory::TitleCategory,
+		TextureCategory::TitleCategory,
 		(int)TitleCategoryTextureList::Background));
 }
